+++ backend/include/supereight/backend/fields/sdf.hpp
@@ -38,7 +38,7 @@ struct voxel_traits<SDF> {
     static value_type initValue() { return {1.f, 0.f}; }
 
     template<typename OctreeT, typename HashType>
-    static int buildAllocationList(HashType* allocation_list, size_t reserved,
+    static int buildAllocationList(HashType* allocation_list, int reserved,
         const OctreeT& octree, const Eigen::Matrix4f& pose,
         const Eigen::Matrix4f& K, const float* depth_map,
         const Eigen::Vector2i& image_size, float mu);
